+++ certbot_dns_ecenter/__init__.py
@@ -1,5 +1,5 @@
 """
-The `certbot_dns_gcore` plugin automates the process of
+The `certbot_dns_ecenter` plugin automates the process of
 completing a ``dns-01`` challenge (`acme.challenges.DNS01`) by creating, and
 subsequently removing, TXT records using the G-Core DNS API.
 
@@ -8,15 +8,15 @@ subsequently removing, TXT records using the G-Core DNS API.
 
    .. code-block:: bash
 
-      pip install certbot-dns-gcore
+      pip install certbot-dns-ecenter
 
 
 Named Arguments
 ---------------
 
 ========================================  =====================================
-``--dns-gcore-credentials``               G-Core credentials INI file. (Required)
-``--dns-gcore-propagation-seconds``       The number of seconds to wait for DNS
+``--dns-ecenter-credentials``                  Edge-Center credentials INI file. (Required)
+``--dns-ecenter-propagation-seconds``          The number of seconds to wait for DNS
                                           to propagate before asking the ACME
                                           server to verify the DNS record. (Default: 10)
 ========================================  =====================================
@@ -25,40 +25,40 @@ Named Arguments
 Credentials
 -----------
 
-Use of this plugin requires a configuration file containing G-Core DNS API
+Use of this plugin requires a configuration file containing Edge-Center DNS API
 credentials.
-You can use G-Core API Token, obtained from your
-`profile panel <https://accounts.gcorelabs.com/profile/api-tokens>`_
+You can use Edge-Center API Token, obtained from your
+`profile panel <https://accounts.edgecenter.ru/profile/api-tokens>`_
 or
-use G-Core Authentication credentials (email and password)
-for `login <https://auth.gcorelabs.com/login/signin>`_ page.
+use Edge-Center Authentication credentials (email and password)
+for `login <https://auth.edgecenter.ru/login/signin>`_ page.
 
-G-Core API Token is recommended authentication option.
+Edge-Center API Token is recommended authentication option.
 
 The token needed by Certbot for add temporary TXT record to zone what you need certificates for.
 
 .. code-block:: ini
    :name: token
-   :caption: Example `gcore.ini` file using restricted API Token (recommended):
+   :caption: Example `ecenter.ini` file using restricted API Token (recommended):
 
-   # G-Core API token used by Certbot
-   dns_gcore_apitoken = 0123456789abcdef0123456789abcdef01234567
+   # Edge-Center API token used by Certbot
+   dns_ecenter_apitoken = 0123456789abcdef0123456789abcdef01234567
 
 .. code-block:: ini
    :name: auth
    :caption: Example `gcore.ini` file using authentication credentials (not recommended):
 
-   # G-Core API credentials used by Certbot
-   dns_gcore_email = gcore_user@example.com
-   dns_gcore_password = 0123456789abcdef0123456789abcdef01234
+   # Edge-Center API credentials used by Certbot
+   dns_ecenter_email = ec_user@example.com
+   dns_ecenter_password = 0123456789abcdef0123456789abcdef01234
 
 The path to this file can be provided interactively or using the
-``--dns-gcore-credentials`` command-line argument. Certbot records the path
+``--dns-ecenter-credentials`` command-line argument. Certbot records the path
 to this file for use during renewal, but does not store the file's contents.
 
 .. caution::
    You should protect these API credentials as you would the password to your
-   G-Core account. Users who can read this file can use these credentials
+   Edge-Center account. Users who can read this file can use these credentials
    to issue arbitrary API calls on your behalf. Users who can cause Certbot to
    run using these credentials can complete a ``dns-01`` challenge to acquire
    new certificates or revoke existing certificates for associated domains,
@@ -71,15 +71,15 @@ file. This warning will be emitted each time Certbot uses the credentials file,
 including for renewal, and cannot be silenced except by addressing the issue
 (e.g., by using a command like ``chmod 600`` to restrict access to the file).
 
-Also you can specify the G-Core ``auth`` and ``dns`` urls.
+Also you can specify the Edge-Center ``auth`` and ``dns`` urls.
 
 .. code-block:: ini
    :name: urls
-   :caption: Example `gcore.ini` file using alternative `auth` and `dns` urls.
+   :caption: Example `ecenter.ini` file using alternative `auth` and `dns` urls.
 
-    # G-Core API credentials used by Certbot
-    dns_gcore_auth_url = https://auth.example.com
-    dns_gcore_api_url = https://dns_api.example.com
+    # Edge-Center API credentials used by Certbot
+    dns_ecenter_auth_url = https://auth.example.com
+    dns_ecenter_api_url = https://dns_api.example.com
 
 Examples
 --------
@@ -88,8 +88,8 @@ Examples
    :caption: To acquire a certificate for ``example.com``
 
    certbot certonly \\
-     --authenticator dns-gcore \\
-     --dns-gcore-credentials=./gcore.ini \\
+     --authenticator dns-ecenter \\
+     --dns-ecenter-credentials=./ecenter.ini \\
      -d 'example.com'
 
 .. code-block:: bash
@@ -97,8 +97,8 @@ Examples
              for DNS propagation
 
    certbot certonly \\
-     --authenticator dns-gcore \\
-     --dns-gcore-credentials=./gcore.ini \\
-     --dns-gcore-propagation-seconds=60 \\
+     --authenticator dns-ecenter \\
+     --dns-ecenter-credentials=./ecenter.ini \\
+     --dns-ecenter-propagation-seconds=60 \\
      -d 'example.com'
 """