+++ certbot_dns_ecenter/dns_ecenter.py
@@ -1,4 +1,4 @@
-"""DNS Authenticator for G-Core."""
+"""DNS Authenticator for Edge-Center."""
 
 import logging
 from typing import Any
@@ -9,20 +9,20 @@ from certbot import errors
 from certbot.plugins import dns_common
 from certbot.plugins.dns_common import CredentialsConfiguration
 
-from . import api_gcore
-from .api_gcore import GCoreConflictException
+from . import api_ecenter
+from .api_ecenter import ECenterConflictException
 
 logger = logging.getLogger(__name__)
 
 
 class Authenticator(dns_common.DNSAuthenticator):
-    """DNS Authenticator for G-Core
+    """DNS Authenticator for Edge-Center
 
-    This Authenticator uses the G-Core DNS API to fulfill a dns-01 challenge.
+    This Authenticator uses the Edge-Center DNS API to fulfill a dns-01 challenge.
     """
 
-    _docs_url = 'https://apidocs.gcorelabs.com/dns#section/Authentication'
-    description = ('Obtain certificates using a DNS TXT record (if you are using G-Core for '
+    _docs_url = 'https://apidocs.edgecenter.ru/dns#section/Authentication'
+    description = ('Obtain certificates using a DNS TXT record (if you are using Edge-Center for '
                    'DNS).')
     ttl = 300
 
@@ -39,10 +39,10 @@ class Authenticator(dns_common.DNSAuthenticator):
             cls, add: Callable[..., None], default_propagation_seconds: int = 10
     ) -> None:
         super().add_parser_arguments(add, default_propagation_seconds)
-        add('credentials', help='G-Core credentials INI file.')
+        add('credentials', help='Edge-Center credentials INI file.')
 
     def more_info(self) -> str:
-        return 'This plugin configures a DNS TXT record to respond to a dns-01 challenge using the G-Core API.'
+        return 'This plugin configures a DNS TXT record to respond to a dns-01 challenge using the Edge-Center API.'
 
     def _validate_credentials(self, credentials: CredentialsConfiguration) -> None:
         self.token = credentials.conf('apitoken')
@@ -54,29 +54,29 @@ class Authenticator(dns_common.DNSAuthenticator):
 
         if self.token:
             if self.email or self.password:
-                raise errors.PluginError('{}: dns_gcore_email and dns_gcore_password are '
+                raise errors.PluginError('{}: dns_ecenter_email and dns_ecenter_password are '
                                          'not needed when using an API Token'
                                          .format(credentials.confobj.filename))
         elif self.email or self.password:
             if not self.email:
-                raise errors.PluginError('{}: dns_gcore_email is required when using a Global '
+                raise errors.PluginError('{}: dns_ecenter_email is required when using a Global '
                                          'API Key. (should be email address associated with '
-                                         'G-Core account)'.format(credentials.confobj.filename))
+                                         'Edge-Center account)'.format(credentials.confobj.filename))
             if not self.password:
-                raise errors.PluginError('{}: dns_gcore_password is required when using a '
+                raise errors.PluginError('{}: dns_ecenter_password is required when using a '
                                          'Global API Key. (see {})'
                                          .format(credentials.confobj.filename, self._docs_url))
         else:
             raise errors.PluginError(
-                '{}: Either dns_gcore_apitoken (recommended), or '
-                'dns_gcore_email and dns_gcore_password are required.'
+                '{}: Either dns_ecenter_apitoken (recommended), or '
+                'dns_ecenter_email and dns_ecenter_password are required.'
                 ' (see {})'.format(credentials.confobj.filename, self._docs_url)
             )
 
     def _setup_credentials(self) -> None:
         self.credentials = self._configure_credentials(
             'credentials',
-            'G-Core credentials INI file',
+            'Edge-Center credentials INI file',
             None,
             self._validate_credentials
         )
@@ -87,17 +87,17 @@ class Authenticator(dns_common.DNSAuthenticator):
     def _cleanup(self, domain: str, validation_name: str, validation: str) -> None:
         self._get_client().del_txt_record(domain, validation_name)
 
-    def _get_client(self) -> "_GCoreClient":
+    def _get_client(self) -> "_ECenterClient":
         if not self.credentials:  # pragma: no cover
             raise errors.Error("Plugin has not been prepared.")
         if self.token:
-            return _GCoreClient(
+            return _ECenterClient(
                 token=self.token,
                 api_url=self.api_url,
                 dns_api_url=self.dns_api_url,
                 auth_url=self.auth_url
             )
-        return _GCoreClient(
+        return _ECenterClient(
             login=self.email,
             password=self.password,
             api_url=self.api_url,
@@ -106,15 +106,15 @@ class Authenticator(dns_common.DNSAuthenticator):
         )
 
 
-class _GCoreClient:
+class _ECenterClient:
     """
-    G-Core client.
+    Edge-Center client.
     """
 
     record_type = 'TXT'
 
     def __init__(self, *args, **kwargs) -> None:
-        self.gcore = api_gcore.GCoreClient(*args, **kwargs)
+        self.gcore = api_ecenter.ECenterClient(*args, **kwargs)
 
     def add_txt_record(
             self, domain: str, record_name: str, record_content: str, record_ttl: int
@@ -126,14 +126,14 @@ class _GCoreClient:
         :param str record_name: The record name (typically beginning with '_acme-challenge.').
         :param str record_content: The record content (typically the challenge validation).
         :param int record_ttl: The record TTL (number of seconds that the record may be cached).
-        :raises certbot.errors.PluginError: if an error occurs communicating with the G-Core DNS API
+        :raises certbot.errors.PluginError: if an error occurs communicating with the Edge-Center DNS API
         """
         domain = self._find_zone_name(domain=domain)
         try:
             self.gcore.record_create(
                 domain, record_name, self.record_type, data=self._data_for_txt(record_ttl, [record_content]),
             )
-        except GCoreConflictException:
+        except ECenterConflictException:
             logger.debug('Record already present on zone. Try to update record content')
             exist_record_content = self.gcore.record_content(domain, record_name, self.record_type)
             if record_content not in exist_record_content:
@@ -157,7 +157,7 @@ class _GCoreClient:
         try:
             domain = self._find_zone_name(domain)
             self.gcore.record_get(domain, record_name, self.record_type)
-        except (api_gcore.GCoreNotFoundException, api_gcore.GCoreConflictException) as err:
+        except (api_ecenter.ECenterNotFoundException, api_ecenter.ECenterConflictException) as err:
             logger.debug('Encountered error finding zone_id during deletion: %s', err)
             return
         self.gcore.record_delete(domain, record_name, self.record_type)
@@ -192,5 +192,5 @@ class _GCoreClient:
             'Unable to determine zone name for {0} using zone names: '
             '{1}. Please confirm that the domain name has been '
             'entered correctly and is already associated with the '
-            'supplied G-Core account.'.format(domain, zone_name_guesses)
+            'supplied Edge-Center account.'.format(domain, zone_name_guesses)
         )