+++ tests/conftest.py
@@ -3,7 +3,7 @@ import responses
 from responses import matchers
 import json
 
-from certbot_dns_gcore.api_gcore import GCoreClient
+from certbot_dns_ecenter.api_ecenter import ECenterClient
 
 
 @pytest.fixture
@@ -51,7 +51,7 @@ def mock_auth():
     # auth
     responses.add(
         responses.POST,
-        f'{GCoreClient._auth_url}/auth/jwt/login',
+        f'{ECenterClient._auth_url}/auth/jwt/login',
         json={'access': 'token'},
         status=200
     )
@@ -62,7 +62,7 @@ def mock_get_zone(record_payload):
     # check domain
     responses.add(
         responses.GET,
-        f'{GCoreClient._dns_api_url}/{GCoreClient._root_zones}/{record_payload["domain"]}',
+        f'{ECenterClient._dns_api_url}/{ECenterClient._root_zones}/{record_payload["domain"]}',
         json={'name': record_payload["domain"]},
         status=200
     )
@@ -73,7 +73,7 @@ def mock_get_zones(record_payload):
     params = {'limit': '100', 'name': domain_2_level}
     responses.add(
         responses.GET,
-        f'{GCoreClient._dns_api_url}/{GCoreClient._root_zones}',
+        f'{ECenterClient._dns_api_url}/{ECenterClient._root_zones}',
         match=[responses.matchers.query_param_matcher(params)],
         json={'zones': [{'name': record_payload["domain"]}]},
         status=200
@@ -84,7 +84,7 @@ def mock_del_record(record_payload):
     # del
     responses.add(
         responses.DELETE,
-        f'{GCoreClient._dns_api_url}/{GCoreClient._root_zones}/{record_payload["domain"]}/{record_payload["record_name"]}/TXT',
+        f'{ECenterClient._dns_api_url}/{ECenterClient._root_zones}/{record_payload["domain"]}/{record_payload["record_name"]}/TXT',
         json={}, status=200
     )
 
@@ -94,7 +94,7 @@ def mock_get_record(record_payload):
     # check record
     responses.add(
         responses.GET,
-        f'{GCoreClient._dns_api_url}/{GCoreClient._root_zones}/{record_payload["domain"]}/{record_payload["record_name"]}/TXT',
+        f'{ECenterClient._dns_api_url}/{ECenterClient._root_zones}/{record_payload["domain"]}/{record_payload["record_name"]}/TXT',
         json={}, status=200
     )
 
@@ -104,7 +104,7 @@ def mock_post_record(record_payload):
     # check record
     responses.add(
         responses.POST,
-        f'{GCoreClient._dns_api_url}/{GCoreClient._root_zones}/{record_payload["domain"]}/{record_payload["record_name"]}/TXT',
+        f'{ECenterClient._dns_api_url}/{ECenterClient._root_zones}/{record_payload["domain"]}/{record_payload["record_name"]}/TXT',
         json={},
         status=200
     )
@@ -114,19 +114,19 @@ def mock_post_record(record_payload):
 def mock_dns_api(record_payload, rrset_exists_two_records, mock_get_zones):
     responses.add(
         responses.POST,
-        f'{GCoreClient._dns_api_url}/{GCoreClient._root_zones}/{record_payload["domain"]}/{record_payload["record_name"]}/TXT',
+        f'{ECenterClient._dns_api_url}/{ECenterClient._root_zones}/{record_payload["domain"]}/{record_payload["record_name"]}/TXT',
         json={},
         status=409
     )
     responses.add(
         responses.GET,
-        f'{GCoreClient._dns_api_url}/{GCoreClient._root_zones}/{record_payload["domain"]}/{record_payload["record_name"]}/TXT',
+        f'{ECenterClient._dns_api_url}/{ECenterClient._root_zones}/{record_payload["domain"]}/{record_payload["record_name"]}/TXT',
         json=json.loads(rrset_exists_two_records),
         status=200
     )
     responses.add(
         responses.PUT,
-        f'{GCoreClient._dns_api_url}/{GCoreClient._root_zones}/{record_payload["domain"]}/{record_payload["record_name"]}/TXT',
+        f'{ECenterClient._dns_api_url}/{ECenterClient._root_zones}/{record_payload["domain"]}/{record_payload["record_name"]}/TXT',
         json={},
         status=200
     )